requirements.txt
annotated-doc>=0.0.4
annotated-types>=0.7.0
anyio>=4.12.1
asyncpg>=0.31.0
argon2-cffi>=23.1.0
bcrypt>=4.2.0
certifi>=2026.2.25
cffi>=2.0.0
charset-normalizer>=3.4.6
click>=8.3.1
ecdsa>=0.19.1
fastapi>=0.135.1
h11>=0.16.0
idna>=3.11
pip>=26.0.1
pyasn1>=0.6.2
pycparser>=3.0
pydantic>=2.12.5
pydantic_core>=2.41.5
python-dotenv>=1.2.2
python-jose>=3.5.0
requests>=2.32.5
rsa>=4.9.1
six>=1.17.0
starlette>=0.52.1
typing_extensions>=4.15.0
typing-inspection>=0.4.2
urllib3>=2.6.3
uvicorn>=0.41.0